Create `MagicRelationsTrait`
