+++ src/ActiveRecord.php
@@ -10,6 +10,7 @@ use Throwable;
 use Yiisoft\ActiveRecord\Trait\ArrayableTrait;
 use Yiisoft\ActiveRecord\Trait\ArrayAccessTrait;
 use Yiisoft\ActiveRecord\Trait\ArrayIteratorTrait;
+use Yiisoft\ActiveRecord\Trait\MagicRelationsTrait;
 use Yiisoft\Arrays\ArrayableInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
@@ -95,6 +96,7 @@ class ActiveRecord extends BaseActiveRecord implements ArrayableInterface, Array
     use ArrayableTrait;
     use ArrayAccessTrait;
     use ArrayIteratorTrait;
+    use MagicRelationsTrait;
 
     /**
      * The insert operation. This is mainly used when overriding {@see transactions()} to specify which operations are
