+++ src/BaseActiveRecordTrait.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\ActiveRecord;
 
-use Error;
 use Exception;
 use ReflectionException;
-use ReflectionMethod;
 use Throwable;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidCallException;
@@ -15,10 +13,8 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\UnknownPropertyException;
 
 use function array_key_exists;
-use function lcfirst;
 use function method_exists;
 use function property_exists;
-use function substr;
 use function ucfirst;
 
 trait BaseActiveRecordTrait
@@ -78,66 +74,6 @@ trait BaseActiveRecordTrait
         }
 
         throw new UnknownPropertyException('Getting unknown property: ' . static::class . '::' . $name);
-    }
-
-    /**
-     * Returns the relation object with the specified name.
-     *
-     * A relation is defined by a getter method which returns an {@see ActiveQueryInterface} object.
-     *
-     * It can be declared in either the Active Record class itself or one of its behaviors.
-     *
-     * @param string $name the relation name, e.g. `orders` for a relation defined via `getOrders()` method
-     * (case-sensitive).
-     * @param bool $throwException whether to throw exception if the relation does not exist.
-     *
-     * @throws InvalidArgumentException if the named relation does not exist.
-     * @throws ReflectionException
-     *
-     * @return ActiveQueryInterface|null the relational query object. If the relation does not exist and
-     * `$throwException` is `false`, `null` will be returned.
-     */
-    public function getRelation(string $name, bool $throwException = true): ActiveQueryInterface|null
-    {
-        $getter = 'get' . ucfirst($name);
-
-        try {
-            /** the relation could be defined in a behavior */
-            $relation = $this->$getter();
-        } catch (Error) {
-            if ($throwException) {
-                throw new InvalidArgumentException(static::class . ' has no relation named "' . $name . '".');
-            }
-
-            return null;
-        }
-
-        if (!$relation instanceof ActiveQueryInterface) {
-            if ($throwException) {
-                throw new InvalidArgumentException(static::class . ' has no relation named "' . $name . '".');
-            }
-
-            return null;
-        }
-
-        if (method_exists($this, $getter)) {
-            /** relation name is case sensitive, trying to validate it when the relation is defined within this class */
-            $method = new ReflectionMethod($this, $getter);
-            $realName = lcfirst(substr($method->getName(), 3));
-
-            if ($realName !== $name) {
-                if ($throwException) {
-                    throw new InvalidArgumentException(
-                        'Relation names are case sensitive. ' . static::class
-                        . " has a relation named \"$realName\" instead of \"$name\"."
-                    );
-                }
-
-                return null;
-            }
-        }
-
-        return $relation;
     }
 
     /**
